requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.0
Werkzeug==2.3.3
PyPDF2==3.0.1
openpyxl==3.1.2
python-pptx==0.6.21
